+++ src/components/User/Step3.jsx
@@ -1,28 +1,21 @@
 import React from "react";
-import { MobileDatePicker } from "@mui/x-date-pickers/MobileDatePicker";
-import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
-import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
 import RadioGroup from "@mui/material/RadioGroup";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Radio from "@mui/material/Radio";
 import FormControl from "@mui/material/FormControl";
-import TextareaAutosize from "@mui/material/TextareaAutosize";
 import {
     Box,
     Button,
     FormHelperText,
     InputAdornment,
     InputLabel,
+    Typography,
 } from "@mui/material";
 import RequiUi from "./RequiUi.jsx";
 import Input from "@mui/material/Input";
-import dayjs from "dayjs";
 
 const Step3 = ({ nextStep, handleChange, formData, prevStep }) => {
-    // MobileDatePicker에서 날짜 선택 시 값을 업데이트하는 핸들러
-    const handleDateChange = (newValue) => {
-        handleChange({ target: { name: "petBirthday", value: newValue } });
-    };
+    // const navigate = useNavigate();
 
     return (
         <Box
@@ -33,79 +26,47 @@ const Step3 = ({ nextStep, handleChange, formData, prevStep }) => {
             mx="auto"
             gap={2}
         >
-            <LocalizationProvider dateAdapter={AdapterDayjs}>
+            <FormControl variant="standard" fullWidth sx={{ mb: 2 }}>
                 <FormHelperText>
-                    아이의 생일은 언제인가요? <RequiUi />
+                    중성화 여부를 알려주세요 <RequiUi />
                 </FormHelperText>
-                <MobileDatePicker
-                    value={dayjs(formData.petBirthday)}
-                    onChange={handleDateChange}
-                />
-            </LocalizationProvider>
+                <RadioGroup
+                    row
+                    aria-required
+                    id="petNeutered"
+                    name="petNeutered"
+                    justifyContent="center"
+                    value={formData.petNeutered}
+                    onChange={handleChange}
+                >
+                    <FormControlLabel value="Y" control={<Radio />} label="O" />
+                    <FormControlLabel value="N" control={<Radio />} label="X" />
+                </RadioGroup>
+            </FormControl>
 
             <FormControl variant="standard" fullWidth sx={{ mb: 4 }}>
-                <InputLabel htmlFor="petWeigth" sx={{ mb: 4 }}>
-                    몸무게를 입력해 주세요 <RequiUi />
+                <InputLabel htmlFor="petFavorite" sx={{ mb: 4 }}>
+                    좋아하는 것을 알려주세요
                 </InputLabel>
                 <Input
-                    required
-                    id="petWeight"
-                    name="petWeight"
-                    placeholder="몸무게를 입력해 주세요"
-                    value={formData.petWeight}
+                    id="petFavorite"
+                    name="petFavorite"
+                    placeholder="ex) 공놀이, 산책"
+                    value={formData.petFavorite}
                     onChange={handleChange}
                     startAdornment={<InputAdornment />}
                 />
             </FormControl>
 
-            <FormControl variant="standard" fullWidth sx={{ mb: 2 }}>
-                <FormHelperText>
-                    체형을 선택해 주세요 <RequiUi />
-                </FormHelperText>
-                <RadioGroup
-                    row
-                    aria-required
-                    id="petBodyType"
-                    name="petBodyType"
-                    justifyContent="center"
-                    value={formData.petBodyType}
-                    onChange={handleChange}
-                >
-                    <FormControlLabel
-                        value="날씬"
-                        control={<Radio />}
-                        label="날씬"
-                    />
-                    <FormControlLabel
-                        value="적당"
-                        control={<Radio />}
-                        label="적당"
-                    />
-                    <FormControlLabel
-                        value="통통"
-                        control={<Radio />}
-                        label="통통"
-                    />
-                </RadioGroup>
-            </FormControl>
+            <Typography variant="body1" mt={3} mb={2}>
+                아이 사진등록하기
+            </Typography>
 
             <FormHelperText>
-                아이를 소개해 주세요 <RequiUi />
+                첫번째 사진으로 프로필 사진이 등록됩니다 <RequiUi />
+                <br />
+                (구현 예정)
             </FormHelperText>
-            <TextareaAutosize
-                id="petIntroduction"
-                name="petIntroduction"
-                minRows={3}
-                placeholder="아이의 특징을 적어주세요"
-                style={{
-                    width: "100%",
-                    padding: "10px",
-                    borderRadius: "5px",
-                    borderColor: "#ccc",
-                }}
-                value={formData.petIntroduction}
-                onChange={handleChange}
-            />
 
             <Button
                 variant="contained"
